+++ lib/common/exception/non_loggable_error.dart
@@ -1,0 +1,1 @@
+abstract class NonLoggableError implements Exception {}