+++ lib/features/drawer/presentation/widgets/account_badge.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:lgpdjus/features/user/account/domain/user_account_entities.dart';
 
 class AccountBadge extends StatelessWidget {
   AccountBadge(this.status);
 
-  final AccountStatus status;
+  final String status;
 
   @override
   Widget build(BuildContext context) {
@@ -21,49 +20,13 @@ class AccountBadge extends StatelessWidget {
           bottomRight: badgeRadius,
         ),
       ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(
-            status.icon,
-            size: 12,
-            color: Color(0xBF3C3C3B),
-          ),
-          SizedBox(
-            width: 4,
-          ),
-          Text(
-            status.text,
-            style: textTheme.caption?.copyWith(
-              fontSize: 10,
-              fontWeight: FontWeight.w800,
-            ),
-          ),
-        ],
+      child: Text(
+        status,
+        style: textTheme.caption?.copyWith(
+          fontSize: 10,
+          fontWeight: FontWeight.w800,
+        ),
       ),
     );
-  }
-}
-
-extension AccountStatusUI on AccountStatus {
-  IconData get icon {
-    switch (this) {
-      case AccountStatus.verified:
-        return Icons.done_outlined;
-      case AccountStatus.unverified:
-        return Icons.cancel_outlined;
-      case AccountStatus.processing:
-        return Icons.hourglass_top_outlined;
-    }
-  }
-  String get text {
-    switch (this) {
-      case AccountStatus.verified:
-        return 'Verificado';
-      case AccountStatus.unverified:
-        return 'Não verificado';
-      case AccountStatus.processing:
-        return 'Em verificação';
-    }
   }
 }