+++ lib/features/user/account/domain/exception/profile_not_found_error.dart
@@ -1,0 +1,3 @@
+import 'package:lgpdjus/common/exception/non_loggable_error.dart';
+
+class ProfileNotFountError implements NonLoggableError {}