+++ CHANGELOG.md
@@ -6,13 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Adicionado
 - Integração com gov.br
 - Item de menu "Sobre recursos de acessibilidade"
 - Item de menu "Sobre permissões e contas"
 
-### Removed
+### Removido
 - Item de menu "Termos de Uso"
+- Rotas de verificação da conta
 
 ## [1.2.3] - 2022-05-23